Module iced_web::css

Style your widgets.

Structs

Css

A cascading style sheet.

Enums

Rule

A CSS rule of a VDOM node.

Functions

alignment

Returns the style value for the given Alignment.

background

Returns the style value for the given Background.

color

Returns the style value for the given Color.

length

Returns the style value for the given Length.

max_length

Returns the style value for the given maximum length in units.

min_length

Returns the style value for the given minimum length in units.

padding

Returns the style value for the given Padding.