Module iced_web::widget

Use the built-in widgets or create your own.

Custom widgets

If you want to implement a custom widget, you simply need to implement the Widget trait. You can use the API of the built-in widgets as a guide or source of inspiration.

Re-exports

For convenience, the contents of this module are available at the root module. Therefore, you can directly type:

use iced_web::{button, Button, Widget};

Re-exports

pub use button::Button;

pub use scrollable::Scrollable;

pub use slider::Slider;

pub use text::Text;

pub use text_input::TextInput;

pub use toggler::Toggler;

pub use checkbox::Checkbox;

pub use container::Container;

pub use image::Image;

pub use progress_bar::ProgressBar;

pub use radio::Radio;

Modules

button

Allow your users to perform actions by pressing a button.

checkbox

Show toggle controls using checkboxes.

container

Decorate content and apply alignment.

image

Display images in your user interface.

progress_bar

Provide progress feedback to your users.

radio

Create choices using radio buttons.

scrollable

Navigate an endless amount of content with a scrollbar.

slider

Display an interactive selector of a single value from a range of values.

text_input

Display fields that can be filled with text.

toggler

Show toggle controls using togglers.

Structs

Column

A container that distributes its contents vertically.

Row

A container that distributes its contents horizontally.

Space

An amount of empty space.

Traits

Widget

A component that displays information and allows interaction.