Struct iced_web::widget::Column

pub struct Column<'a, Message> { /* fields omitted */ }

A container that distributes its contents vertically.

A Column will try to fill the horizontal space of its container.

Implementations

impl<'a, Message> Column<'a, Message>

pub fn new() -> Self

Creates an empty Column.

pub fn with_children(children: Vec<Element<'a, Message>>) -> Self

Creates a Column with the given elements.

pub fn spacing(self, units: u16) -> Self

Sets the vertical spacing between elements.

Custom margins per element do not exist in Iced. You should use this method instead! While less flexible, it helps you keep spacing between elements consistent.

pub fn padding<P: Into<Padding>>(self, padding: P) -> Self

Sets the Padding of the Column.

pub fn width(self, width: Length) -> Self

Sets the width of the Column.

pub fn height(self, height: Length) -> Self

Sets the height of the Column.

pub fn max_width(self, max_width: u32) -> Self

Sets the maximum width of the Column.

pub fn max_height(self, max_height: u32) -> Self

Sets the maximum height of the Column in pixels.

pub fn align_items(self, align: Alignment) -> Self

Sets the horizontal alignment of the contents of the Column .

pub fn push<E>(self, child: E) -> Self where
E: Into<Element<'a, Message>>,

Adds an element to the Column.